\name{cat}
\alias{cat}
\title{ modified version of cat for relax }
\description{Output of objects, see help of \code{base::cat} }
\usage{
cat(..., file = "", sep = " ", fill = FALSE, labels = NULL, append = FALSE)
}
\arguments{
  \item{\dots}{ further arguments }
  \item{file}{ file name }
  \item{sep}{ separator }
  \item{fill}{ for controlling line breaks }
  \item{labels}{ to label lines printed }
  \item{append}{ to append infos to a file}
}
\details{ see help of cat (package base) }
\value{ invisible NULL }
\seealso{ \code{\link{cat}} }
\examples{
##
}
\keyword{ IO }

