\name{menu}
\alias{menu}
\title{modified version of menu for relax }
\description{ \code{menu} presents user some choices for choosing }
\usage{ menu(choices, graphics = FALSE, title = "") }
\arguments{
  \item{choices}{ vector of choices }
  \item{graphics}{ whether a graphics menu should be used }
  \item{title}{ used as title }
}
\details{ see help of menu (package utils) }
\value{ number of selected item }
\seealso{ \code{\link{menu}} }
\examples{ ## }
\keyword{ IO }

