\name{print}
\alias{print}
\title{ modified version of print for relax }
\description{ \code{print} prints its argument }
\usage{ print(x, ...) }
\arguments{
  \item{x}{ object to be printed }
  \item{\dots}{ further arguments }
}
\details{ see help of print (package base) }
\seealso{ \code{\link{print}} }
\examples{ ## }
\keyword{ IO }

