\name{readline}
\alias{readline}
\title{modified version of readline for relax }
\description{ reads a line from the terminal }
\usage{ readline(prompt = "") }
\arguments{
  \item{prompt}{ string printed }
}
\details{ see help of readline (package base) }
\seealso{ \code{\link{readline}} }
\examples{ ## }
\keyword{ IO }

