\name{str}
\alias{str}
\title{ modified version of str for relax }
\description{  compactly display the structure of an arbitrary R object, see help of \code{utils::str}}
\usage{ str(object, ...) }
\arguments{ 
  \item{object}{ R object }
  \item{\dots}{ further arguments }
}
\details{ see help of str (package utils) }
\seealso{ \code{\link{str}} }
\examples{
##
}
\keyword{ IO }
